adapt driver and CBS to changed heuristic code
I split CBSHeuristic into multiple classes to avoid large code chunk in each
function.

I test in another repo (which I might not push to remote), where I compare the
h-val of new heuristic code to that of the old code side by side and seems they
are the same.